1차 merge

+++ src/main/java/Albaid/backend/domain/contract/repository/ContractRepository.java
@@ -6,10 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-import java.util.List;
-
 public interface ContractRepository extends JpaRepository<Contract, Integer> {
-    List<Contract> findByMemberId(Long memberId);
+    List<Contract> findByMemberId(Integer memberId);
 
 
     @Query("SELECT c FROM Contract c WHERE c.member.id = :memberId")
